pyproject.toml
[project]
name = "create-avatar"
version = "0.1.0"
description = "Offline avatar generation pipeline: photo/parametric → GLB with ARKit blendshapes + visemes"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "mediapipe>=0.10.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "scikit-image>=0.22.0",
    "trimesh>=4.0.0",
    "Pillow>=10.0.0",
    "pillow-heif>=0.15.0",
    "pygltflib>=1.16.0",
    "PyYAML>=6.0",
    "click>=8.1.0",
    "yacs>=0.1.8",
    "chumpy>=0.70",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff"]

[project.scripts]
create-avatar = "create_avatar.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
